Serious speed improvements
The use of a package not in the standard library allowed the creation of the degenerate primer regex much faster because it does not produce duplicate permutations. Additionally, cycling through each possible primer from the least number of errors to the greatest saves time since most primers are found with fewer than the amount of errors allowed.